eb_infex_worker/information_extraction/requirements.txt
cycler==0.10.0
decorator==5.0.7
docopt==0.6.2
fake-useragent==1.1.3
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.3.3
networkx==2.5
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
pandas==2.0.1
parsel==1.6.0
Pillow==9.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2022.6.2
scikit-learn==1.1.3
scipy==1.5.4
selenium==4.5.0
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.56.0
timeout-decorator==0.5.0
tldextract==3.4.4
urllib3==1.26.12
xgboost==1.4.2
ImageHash==4.3.1